useReducer-practice
Here I have re-created my date counter application however I have made use of the newest react tool I have added to my toolkit which is useReducer. I have used it to manage the multiple states of count and step and made use of a switch statements to tell the reducer function what it should do with the given input type. This allows me to keep all the component logic nicely together and allows for easier updating of the states. I will add to this as a learn more about the useReducer hook.