useReducer real world practice
Over the last few days i have tried to really drive home the concepts and use cases for useReducer as i feel it is such a powerful tool that will be really handy for me down the line. I have built here a still simple app but much more complex than the DateCounter before it, it is a react quiz application. To get the questions i have simulated an API on my machine with a json file but if needed simply move the questions array into the main file and use it directly or change the fetch URL to an api that returns questions in json, I have used the useReducer to store all the state of this application and handle each dispatch with a switch statement and destructured the state straight away rather than to use state.x all over the place. I have also utilised some more useEffects for things such as the timer and the fetching of the data and a good amount of conditional rendering for certain components such as the loader or end screen. I feel this small project has really driven home the react tools utilised within it and i feel my improvements every day leading to cleaner and more optimised code. I am very excited to continue using this tool in future projects.